refactor(toolkit): collapse form-field public_api into index, refresh headless README
- Inline the contents of `form-field/public_api.ts` into `index.ts` and
  delete the empty re-export shim. `ng-package.json` already points at
  `index.ts`, so the shim was doing nothing useful and doubling the
  cognitive surface for anyone grepping public exports.
- Update the workspace path maps (`tsconfig.base.json`, `apps/demo`,
  `packages/toolkit/vitest.shared.mts`) and the one browser-spec
  deep-import to point at `index.ts`.
- Headless README: expose the `submittedStatus` input in the fieldset
  directive table, list the full resolved signal set (including the
  renamed `resolvedSubmittedStatus`), and update the field-name
  directive blurb so it describes the new null-return + dev-mode
  console.error contract.
